Big database update, Added databaseinterface as an abstraction from the database and its tables, added findplayerstatement, added the possiblity to create new players with keys, keys and get player information though databaseinterface, fixed many bug and warnings, updated databasedevice

+++ Client/source/device/NetDevice.h
@@ -119,7 +119,7 @@ namespace Device::Net
 			const EResult result = networkInterface->SendMessageToConnection(
 				connection,
 				&messagePipe.getData()[0],
-				messagePipe.getSize(),
+				(uint32) messagePipe.getSize(),
 				k_nSteamNetworkingSend_Reliable,
 				NULL);
